requirements.txt
Django==2.2.6
djangorestframework==3.10.3
mysqlclient==1.4.4
pytz==2019.3
sqlparse==0.3.0
xlrd==1.2.0
xlwt==1.3.0
